Password toggle function separated
Password toggle function separated.
+bug fixed
