+++ src/Functions/PasCloseEyeIcon.jsx
@@ -1,0 +1,13 @@
+export function PasCloseEyeIcon() {
+    // toggle the type attribute
+    const togglePassword = document.querySelector("#togglePassword");
+    const passwordV = document.querySelector("#password_field");
+    const type = passwordV.getAttribute("type") === "password" ? "text" : "password";
+
+    togglePassword.className === "fa fa-eye viewpass mr-4 text-muted"
+      ? (document.getElementById("togglePassword").className =
+          "fa fa-eye-slash viewpass mr-4 text-muted")
+      : (document.getElementById("togglePassword").className =
+          "fa fa-eye viewpass mr-4 text-muted");
+    passwordV.setAttribute("type", type);
+  }
